requirements.txt
hikari-lightbulb
python-dotenv
apscheduler